Increse otp digits generated by defaults
The OTP library supports just the minimum of up to 9 digits which is what is
documented in https://www.rfc-editor.org/rfc/rfc4226#section-5.3. While this is perfectly fine
for human usage, also RFC defines that this is an example and more digits should be used for
a in increased security. As our case is not human-facing, we can bump this to the maximum
size allowed by the hashing algorithm.

Credits to @aleksej-paschenko for discoverying that the gotp lib returned 0 padded strings!