Use a Null connection manager
ConnManager needs to exist in the config otherwise if libp2p tries to
access it like here[0], it can panic as its a nil object. In fact, all
the defaults should never be empty as it could lead to accessing nil
references.

This adds the NullManager to the default config it its nil to avoid
panics when libp2p calls functions on it

[0] https://github.com/libp2p/go-libp2p/blob/v0.35.2/config/config.go?rgh-link-date=2024-07-10T08%3A58%3A43Z#L336

Signed-off-by: Itxaka <itxaka@kairos.io>